Optimize particle entanglement entropy
Using the occupation basis instead of the first-quantized particle basis
results in much smaller matrices and faster execution. The particle
entropy calculation is still the bottleneck, but by a much smaller
factor. For example, with 8 sites and 8 particles, the overall
calculation (including diagonalization and spatial entropy) is over 100
times faster.